Extract notebook actions and use them in table view too
