+++ src/components/notebooks/NotebookActions.tsx
@@ -1,0 +1,145 @@
+import { useMutation } from "@tanstack/react-query";
+import { MoreHorizontal, Share2, Trash2 } from "lucide-react";
+import React, { useState } from "react";
+import { useTrpc } from "../TrpcProvider";
+import { Button } from "../ui/button";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+} from "../ui/dialog";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "../ui/dropdown-menu";
+import { SharingModal } from "./SharingModal";
+import type { NotebookProcessed } from "./types";
+
+interface NotebookActionsProps {
+  notebook: NotebookProcessed;
+  onUpdate?: () => void;
+  className?: string;
+}
+
+export const NotebookActions: React.FC<NotebookActionsProps> = ({
+  notebook,
+  onUpdate,
+  className = "",
+}) => {
+  const [isSharingModalOpen, setIsSharingModalOpen] = useState(false);
+  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
+  const trpc = useTrpc();
+
+  // Delete notebook mutation
+  const deleteNotebookMutation = useMutation(
+    trpc.deleteNotebook.mutationOptions()
+  );
+
+  const canEdit = notebook.myPermission === "OWNER";
+
+  const handleShare = (e: React.MouseEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+    setIsSharingModalOpen(true);
+  };
+
+  const handleDeleteClick = (e: React.MouseEvent) => {
+    e.preventDefault();
+    e.stopPropagation();
+    setIsDeleteDialogOpen(true);
+  };
+
+  const handleDeleteConfirm = async () => {
+    try {
+      await deleteNotebookMutation.mutateAsync({
+        nbId: notebook.id,
+      });
+
+      // Call onUpdate to refresh the notebook list
+      onUpdate?.();
+      setIsDeleteDialogOpen(false);
+    } catch (error) {
+      console.error("Failed to delete notebook:", error);
+      // TODO: Show error toast
+    }
+  };
+
+  if (!canEdit) {
+    return null;
+  }
+
+  return (
+    <>
+      <DropdownMenu>
+        <DropdownMenuTrigger asChild>
+          <Button
+            variant="ghost"
+            size="sm"
+            className={`h-8 w-8 p-0 ${className}`}
+            onClick={(e) => {
+              e.preventDefault();
+              e.stopPropagation();
+            }}
+          >
+            <MoreHorizontal className="h-4 w-4" />
+          </Button>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end">
+          <DropdownMenuItem onClick={handleShare}>
+            <Share2 className="mr-2 h-4 w-4" />
+            Share
+          </DropdownMenuItem>
+          <DropdownMenuItem
+            onClick={handleDeleteClick}
+            className="text-red-600 focus:text-red-600"
+          >
+            <Trash2 className="mr-2 h-4 w-4" />
+            Delete
+          </DropdownMenuItem>
+        </DropdownMenuContent>
+      </DropdownMenu>
+
+      {/* Sharing Modal */}
+      <SharingModal
+        notebook={notebook}
+        isOpen={isSharingModalOpen}
+        onClose={() => setIsSharingModalOpen(false)}
+        onUpdate={onUpdate}
+      />
+
+      {/* Delete Confirmation Dialog */}
+      <Dialog open={isDeleteDialogOpen} onOpenChange={setIsDeleteDialogOpen}>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>Delete Notebook</DialogTitle>
+            <DialogDescription>
+              Are you sure you want to delete "
+              {notebook.title || "Untitled Notebook"}"? This action cannot be
+              undone and will permanently remove the notebook and all its
+              contents.
+            </DialogDescription>
+          </DialogHeader>
+          <div className="flex justify-end gap-2 pt-4">
+            <Button
+              variant="outline"
+              onClick={() => setIsDeleteDialogOpen(false)}
+            >
+              Cancel
+            </Button>
+            <Button
+              onClick={handleDeleteConfirm}
+              className="bg-red-600 hover:bg-red-700 focus:ring-red-600"
+              disabled={deleteNotebookMutation.isPending}
+            >
+              {deleteNotebookMutation.isPending ? "Deleting..." : "Delete"}
+            </Button>
+          </div>
+        </DialogContent>
+      </Dialog>
+    </>
+  );
+};
